+++ AVWXKit/Network/AVWXClient.swift
@@ -21,75 +21,6 @@ public struct AVWXClient {
         case network(Error)
     }
 
-    public struct MetarOptions: OptionSet {
-        public let rawValue: Int
-        
-        public init(rawValue: Int) {
-            self.rawValue = rawValue
-        }
-        
-        public static let speech = MetarOptions(rawValue: 1 << 0)
-        public static let info = MetarOptions(rawValue: 1 << 1)
-        public static let translate = MetarOptions(rawValue: 1 << 2)
-        
-        func string() -> String? {
-            var result = [String]()
-            if self.isEmpty {
-                return nil
-            }
-            if self.contains(.info) {
-                result.append("info")
-            }
-            if self.contains(.translate) {
-                result.append("translate")
-            }
-            if self.contains(.speech) {
-                result.append("speech")
-            }
-            
-            return result.joined(separator: ",")
-        }
-    }
-
-    enum Endpoint {
-        
-        case metar(String, MetarOptions)
-        case metarCoordintes(CLLocationCoordinate2D, MetarOptions)
-        case taf(String)
-        
-        func url(baseURL: URL) -> URL {
-            switch self {
-            case .metar(let icao, let options):
-                var fullUrl = baseURL.appendingPathComponent("metar/\(icao)")
-                fullUrl = Endpoint.addOptions(options: options, to: fullUrl)
-                return fullUrl
-                
-            case .metarCoordintes(let coordinates, let options):
-                let formattedLat = String(format: "%0.3f", coordinates.latitude)
-                let formattedLong = String(format: "%0.3f", coordinates.longitude)
-                var fullUrl = baseURL.appendingPathComponent("metar/\(formattedLat),\(formattedLong)")
-                fullUrl = Endpoint.addOptions(options: options, to: fullUrl)
-                return fullUrl
-                
-            case .taf(let icao):
-                return baseURL.appendingPathComponent("taf/\(icao)")
-            }
-        }
-        private static func addOptions(options: MetarOptions, to url: URL) -> URL {
-            var resultUrl = url
-            if
-                let optionValues = options.string(),
-                var components = URLComponents(url: url, resolvingAgainstBaseURL: false)
-            {
-                components.queryItems =  [URLQueryItem(name: "options", value: optionValues)]
-                if let url = components.url {
-                    resultUrl = url
-                }
-            }
-            return resultUrl
-        }
-    }
-    
     public init(baseURL: URL = URL(string: "https://avwx.rest/api/")!, token: String) {
         self.baseURL = baseURL
         self.token = token
